Move common response DTO to `common`
