+++ src/main/scala/com/hhandoko/realworld/route/AuthRoutes.scala
@@ -3,13 +3,13 @@ package com.hhandoko.realworld.route
 import cats.Applicative
 import cats.effect.{ContextShift, Sync}
 import cats.implicits._
-import io.circe.{Encoder, Json}
 import io.circe.generic.auto._
 import org.http4s.circe.{jsonEncoderOf, jsonOf}
 import org.http4s.dsl.Http4sDsl
 import org.http4s.{EntityEncoder, HttpRoutes}
 
 import com.hhandoko.realworld.auth.{AuthService, UnauthorizedResponseSupport}
+import com.hhandoko.realworld.route.common.UserResponse
 
 object AuthRoutes extends UnauthorizedResponseSupport {
 
@@ -40,22 +40,4 @@ object AuthRoutes extends UnauthorizedResponseSupport {
   }
 
   final case class LoginPostPayload(email: String, password: String)
-
-  // TODO: Duplicated in UserRoutes, consolidate
-  final case class UserResponse(email: String, token: String, username: String, bio: Option[String], image: Option[String])
-
-  object UserResponse {
-    implicit val encoder: Encoder[UserResponse] = (r: UserResponse) => Json.obj(
-      "user" -> Json.obj(
-        "email"    -> Json.fromString(r.email),
-        "token"    -> Json.fromString(r.token),
-        "username" -> Json.fromString(r.username),
-        "bio"      -> r.bio.fold(Json.Null)(Json.fromString),
-        "image"    -> r.image.fold(Json.Null)(Json.fromString)
-      )
-    )
-
-    implicit def entityEncoder[F[_]: Applicative]: EntityEncoder[F, UserResponse] =
-      jsonEncoderOf[F, UserResponse]
-  }
 }
